Allow override:
Two additional options introduced:

* allow_secrets
* disallow_secrets

"disallow_secrets" can be set as the policies we should enforce.
If none are provided by config, we take the defaults we know about,
which are all of them.

Then "allow_secrets" allows to set secrets which are allowed and
will not trigger a violation.

I tried testing this in the test script but I couldn't set the
configuration option. :/

Otherwise, it works. :P