Fix disallow code:
There was a major bug in the code I didn't spot.

Instead of creating a list from the default *OR* the disallow,
we create from the default and then check against the disallow,
and then use the ones we got from the default. In short, buggy.

This is now cleaner:

* Create a list from the disallow or - if that doesn't exist - from
  the default.
* Check that against the allow list on whether to check it or not.

I've also added more tests to verify this.